Send an open file instead of just the path
