+++ src/xlib/gtk.h
@@ -1,0 +1,13 @@
+#include "../util.h"
+
+void ugtk_openfilesend(void);
+
+void ugtk_openfileavatar(void);
+
+void ugtk_native_select_dir_ft(uint32_t fid, MSG_FILE *file);
+
+void ugtk_savefiledata(MSG_FILE *file);
+
+void ugtk_save_chatlog(uint32_t friend_number);
+
+void *ugtk_load(void);
